Create unplur.js
A add-on to an inventory function to catch the instance of a single item and remove the 's' from the plural form of 'label'. Each 'label' string would have to be entered initially in it's plural form which would persist for all other numbers, including '0'. Exceptions would have to be made for anomalies like 'goose - geese', etc, depending on inventory type.